backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# CORS and Security
python-jose[cryptography]==3.3.0

# Data Processing
numpy==1.26.3
pillow==10.2.0
pandas==2.2.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Optional: For production YOLO integration
# ultralytics==8.1.0
# torch==2.1.2
# torchvision==0.16.2

# Utilities
python-dateutil==2.8.2
